refactor: move everything from index.js into a new file library.js
